parity fixes
