+++ src/webapps/single_qubit/single.html
@@ -1,0 +1,157 @@
+<!doctype html>
+
+<html>
+<head>
+    <!-- Recommended meta tags -->
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1.0">
+
+    <!-- PyScript CSS -->
+    <link rel="stylesheet" href="https://pyscript.net/releases/2024.1.1/core.css">
+    <!-- CSS for sidebar -->
+    <style>
+        #sidebar {
+            width: 350px;
+        }
+    </style>
+
+
+    <!-- This script tag bootstraps PyScript -->
+    <script type="module" src="https://pyscript.net/releases/2024.1.1/core.js"></script>
+
+    <script src="https://unpkg.com/@microsoft/quantum-viz.js"></script>
+
+    <!-- for splashscreen -->
+    <style>
+        #loading {
+            outline: none;
+            border: none;
+            background: transparent
+        }
+    </style>
+    <script type="module">
+        const loading = document.getElementById('loading');
+        addEventListener('py:ready', () => loading.close());
+        loading.showModal();
+    </script>
+
+    <title>Hume</title>
+    <link rel="icon" type="image/png" href="./assets/favicon.png"/>
+
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css"/>
+
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" type="text/css"/>
+
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@holoviz/panel@1.3.1/dist/css/widgets.css" type="text/css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@holoviz/panel@1.3.1/dist/css/markdown.css" type="text/css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@holoviz/panel@1.3.1/dist/bundled/bootstraptemplate/bootstrap.css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@holoviz/panel@1.3.1/dist/bundled/theme/default.css"/>
+
+    <!-- bokeh scripts -->
+    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-3.2.2.js"></script>
+    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-widgets-3.2.2.min.js"></script>
+    <script type="text/javascript" src="https://cdn.bokeh.org/bokeh/release/bokeh-tables-3.2.2.min.js"></script>
+    <script type="text/javascript">
+        Bokeh.set_log_level("info");
+    </script>
+
+    <!-- panel scripts -->
+    <script defer src="https://cdn.jsdelivr.net/npm/@holoviz/panel@1.3.8/dist/panel.min.js"></script>
+</head>
+
+<body>
+    <dialog id="loading">
+        <h1>Loading...</h1>
+    </dialog>
+
+    <nav class="navbar" style="background-color: #000000">
+        <div class="app-header">
+            <a href="/">
+                <img src="https://avatars.githubusercontent.com/u/163343503?v=4" width="80" class="logo" />
+            </a>
+            <a class="title" href="" style="color: #f0ab3c">Learn Quantum Computing: Single-Qubit Circuits</a>
+        </div>
+    </nav>
+
+    <section class="pyscript">
+        <div class="row overflow-hidden" id="content">
+            <div class="sidenav" id="sidebar">
+                <ul class="nav flex-column">
+                    <div class="bk-root" id="gate-widget"></div>
+                    <div class="bk-root" id="arg-widget"></div>
+                    <div class="bk-root" id="go-widget"></div>
+                </ul>
+            </div>
+            <div class="col mh-100 float-right" id="main">
+                <div class="bk-root" id="circuit_title"></div>
+                <div class="bk-root" id="circuit"></div>
+                <div class="bk-root" id="info_title"></div>
+                <div class="bk-root" id="info"></div>
+            </div>
+        </div>
+
+        <script type="py" src="./main.py" config="./pyscript.toml"></script>
+    </section>
+
+    <script>
+        function show_state(state) {
+            const stateDiv = document.getElementById('state');
+            if (stateDiv != null) {
+                stateDiv.innerHTML = '<u><b>State</b></u></br>' + '<pre>' + state + '</pre>'
+            }
+        }
+
+        function draw(circ) {
+            circ = eval('(' + circ + ')');
+            const circuitDiv = document.getElementById('circuit');
+            if (circuitDiv != null) {
+                document.getElementById('circuit_title').innerHTML = '<u>Circuit</u>'
+                qviz.draw(circ, circuitDiv, qviz.STYLES['BlackAndWhite']);
+            }
+
+            document.getElementById('info_title').innerHTML = '<u>State</u>'
+        }
+
+    </script>
+
+    <py-script>
+        import js
+
+        gate = pn.widgets.Select(name='Gate', options=[None] + [gate.upper() for gate in gates]).servable(target='gate-widget')
+        arg = pn.widgets.NumberInput(name='Angle (in degrees)', disabled=True).servable(target='arg-widget')
+        go = pn.widgets.Button(name='Apply', button_type='primary').servable(target='go-widget')
+
+        info = pn.pane.Str('').servable(target='info')
+
+        qc = create_single_qubit()
+
+        @pn.depends(gate, watch=True)
+        def select_gate(v):
+            arg.disabled = gate.value is None or not (gate.value.lower() in arg_gates)
+            print('gate', v)
+
+        @pn.depends(go, watch=True)
+        def apply(v):
+            global out
+            if go.value is True and gate.value is not None:
+                # add_gate(qc, [], 0, gate.value.lower(), arg.value/180*pi if gate.value.lower() in arg_gates else None)
+                apply_gate(qc, gate.value.lower(), arg.value)
+                gate.value = None
+                arg.value = None
+                arg.disabled = True
+                go.value = False
+
+                s = get_state(qc)
+
+                qc_str = circuit_to_string(qc) #component.get_circuit()
+                js.draw(qc_str)
+                out = f'Step {last_step(qc)}\n-------\n\n{s}\n\n{out}'
+
+            info.object = out
+
+        reset(qc)
+        out = f'Initial state\n-------------\n\n{get_state(qc)}'
+        info.object = out
+    </py-script>
+</body>
+</html>
